requirements.txt

Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.2.1
#tensorflow==2.15.0
numpy==1.22.3
Pillow==9.0.1
opencv-python==4.8.0.76
email-validator==2.1.0
python-dotenv==1.0.0
werkzeug==2.0.1